requirements.txt
edge-tts>=6.1.9
yfinance>=0.2.40
gTTS
aiofiles
selenium
webdriver-manager
pydub>=0.25.1
pytchat
google-generativeai
openai
python-dotenv
feedparser
requests
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
